requirements.txt
flask
flask_bootstrap
flask_session
flask-mysql
pyyaml